Creating user profile page

+++ server/controllers/user.controller.js
@@ -20,6 +20,12 @@ export const getUser = async (req, res, next) => {
 };
 
 export const updateUser = async (req, res, next) => {
+   if (req.user.id !== req.params.id) {
+      return next({
+         status: 401,
+         message: "You can only update your own account/profile!",
+      });
+   }
    try {
       if (req.body.password) {
          req.body.password = await bcrypt.hash(req.body.password, 10);
@@ -44,6 +50,12 @@ export const updateUser = async (req, res, next) => {
 };
 
 export const deleteUser = async (req, res, next) => {
+   if (req.user.id !== req.params.id) {
+      return next({
+         status: 401,
+         message: "You can only delete your own account/profile!",
+      });
+   }
    try {
       const query = { _id: new ObjectId(req.params.id) };
       await collection.deleteOne(query);
